Refactor main.cpp into different files
